Dejitaru Tsuka Ecosystem White Paper
THE DEJITARU TSUKA ECOSYSTEM is a dynamic and innovative blockchain ecosystem pioneering the development of the Dragon Chain. It is designed to provide a secure, efficient and community-driven platform for decentralized applications (dApps) and digital assets. By design, Dejitaru Tsuka (TSUKA) serves as the primary gas token — akin to ETH on the Ethereum network — driving transaction processing and computational operations. However, the true essence of TSUKA lies in its commitment to community building and collaboration with developers and thousands of dedicated members working hand-in-hand to shape the future of DeFi.